refactor(grpc): rename package and update dependencies
- Rename package from "grpc_demo" to "example/grpc"
- Update import paths for proto files- Remove unused import of "github.com/gone-io/gone/v2"
- Update google.golang.org/grpc to v1.72.0
- Add OpenTelemetry dependencies
