+++ examples/grpc/example/go.mod
@@ -1,4 +1,4 @@
-module grpc_demo
+module example/grpc
 
 go 1.24.1
 
@@ -6,18 +6,20 @@ require (
 	github.com/gone-io/gone/v2 v2.1.0
 	github.com/gone-io/goner/grpc v1.1.1
 	github.com/gone-io/goner/viper v1.1.1
-	google.golang.org/grpc v1.71.1
+	google.golang.org/grpc v1.72.0
 	google.golang.org/protobuf v1.36.6
 )
 
 replace (
-	github.com/gone-io/goner/g => ../../g
-	github.com/gone-io/goner/grpc => ../
-	github.com/gone-io/goner/viper => ../../viper
+	github.com/gone-io/goner/g => ./../../../g
+	github.com/gone-io/goner/grpc => ./../../../grpc
+	github.com/gone-io/goner/viper => ./../../../viper
 )
 
 require (
 	github.com/fsnotify/fsnotify v1.8.0 // indirect
+	github.com/go-logr/logr v1.4.2 // indirect
+	github.com/go-logr/stdr v1.2.2 // indirect
 	github.com/go-viper/encoding/javaproperties v0.1.0 // indirect
 	github.com/go-viper/mapstructure/v2 v2.2.1 // indirect
 	github.com/gone-io/goner/g v1.1.1 // indirect
@@ -30,6 +32,11 @@ require (
 	github.com/spf13/pflag v1.0.6 // indirect
 	github.com/spf13/viper v1.20.1 // indirect
 	github.com/subosito/gotenv v1.6.0 // indirect
+	go.opentelemetry.io/auto/sdk v1.1.0 // indirect
+	go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc v0.60.0 // indirect
+	go.opentelemetry.io/otel v1.35.0 // indirect
+	go.opentelemetry.io/otel/metric v1.35.0 // indirect
+	go.opentelemetry.io/otel/trace v1.35.0 // indirect
 	go.uber.org/mock v0.5.1 // indirect
 	go.uber.org/multierr v1.11.0 // indirect
 	golang.org/x/net v0.38.0 // indirect
