+++ examples/grpc/example/v1_server/main.go
@@ -2,28 +2,27 @@ package main
 
 import (
 	"context"
-	"github.com/gone-io/gone/v2"
+	"example/grpc/proto"
 	goneGrpc "github.com/gone-io/goner/grpc"
 	"google.golang.org/grpc"
-	"grpc_demo/proto"
 	"log"
 	"os"
 )
 
 type server struct {
 	gone.Flag
-	proto.UnimplementedHelloServer              // 嵌入UnimplementedHelloServer
-	grpcServer                     *grpc.Server `gone:"*"` // 注入grpc.Server
+	proto.UnimplementedHelloServer // 嵌入UnimplementedHelloServer
 }
 
-func (s *server) Init() {
-	proto.RegisterHelloServer(s.grpcServer, s) //注册服务
-}
-
-// Say  重载协议中定义的服务
+// 重载协议中定义的服务
 func (s *server) Say(ctx context.Context, in *proto.SayRequest) (*proto.SayResponse, error) {
 	log.Printf("Received: %v", in.GetName())
 	return &proto.SayResponse{Message: "Hello " + in.GetName()}, nil
+}
+
+// 实现 goneGrpc.Service接口的RegisterGrpcServer方法，该方法在服务器启动时会被自动调用
+func (s *server) RegisterGrpcServer(server *grpc.Server) {
+	proto.RegisterHelloServer(server, s)
 }
 
 func main() {
